added direct clone return
