+++ rules-tests/Defluent/Rector/Return_/DefluentReturnMethodCallRector/Fixture/skip_clone_object_assign.php.inc
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rector\Tests\Defluent\Rector\Return_\DefluentReturnMethodCallRector\Fixture;
+
+use Rector\Tests\Defluent\Rector\Return_\DefluentReturnMethodCallRector\Source\SelfButCloneAssignVersion;
+
+final class SkipCloneObjectAssign
+{
+    public function run(SelfButCloneAssignVersion $selfButCloneVersion)
+    {
+        return $selfButCloneVersion->withStatus(500);
+    }
+}
