+++ rules-tests/Defluent/Rector/Return_/DefluentReturnMethodCallRector/Fixture/skip_new_object.php.inc
@@ -2,12 +2,22 @@
 
 namespace Rector\Tests\Defluent\Rector\Return_\DefluentReturnMethodCallRector\Fixture;
 
-use Rector\Tests\Defluent\Rector\Return_\DefluentReturnMethodCallRector\Source\SelfButNewVersion;
+use Rector\Tests\Defluent\Rector\Return_\DefluentReturnMethodCallRector\Source\SelfStaticButNewVersion;
 
 final class SkipNewObject
 {
-    public function run(SelfButNewVersion $selfButNewVersion)
+    public function run(SelfStaticButNewVersion $selfButNewVersion)
     {
         return $selfButNewVersion->duplicate();
+    }
+
+    public function run2(SelfStaticButNewVersion $selfButNewVersion)
+    {
+        return $selfButNewVersion->duplicate2();
+    }
+
+    public function run3(SelfStaticButNewVersion $selfButNewVersion)
+    {
+        return $selfButNewVersion->duplicate3();
     }
 }
